+++ django_safe_migrations/reporters/__init__.py
@@ -3,6 +3,7 @@
 from django_safe_migrations.reporters.base import BaseReporter
 from django_safe_migrations.reporters.console import ConsoleReporter
 from django_safe_migrations.reporters.github import GitHubReporter
+from django_safe_migrations.reporters.github_pr import GitHubPRReporter
 from django_safe_migrations.reporters.gitlab import GitLabReporter
 from django_safe_migrations.reporters.json_reporter import JsonReporter
 from django_safe_migrations.reporters.sarif import SarifReporter
@@ -12,6 +13,7 @@ __all__ = [
     "ConsoleReporter",
     "JsonReporter",
     "GitHubReporter",
+    "GitHubPRReporter",
     "GitLabReporter",
     "SarifReporter",
     "get_reporter",
@@ -22,7 +24,8 @@ def get_reporter(format_name: str, **kwargs: object) -> BaseReporter:
     """Get a reporter instance by format name.
 
     Args:
-        format_name: One of 'console', 'json', 'github', 'gitlab', 'sarif'.
+        format_name: One of 'console', 'json', 'github', 'github-pr',
+            'gitlab', 'sarif'.
         **kwargs: Additional arguments to pass to the reporter.
 
     Returns:
@@ -35,6 +38,7 @@ def get_reporter(format_name: str, **kwargs: object) -> BaseReporter:
         "console": ConsoleReporter,
         "json": JsonReporter,
         "github": GitHubReporter,
+        "github-pr": GitHubPRReporter,
         "gitlab": GitLabReporter,
         "sarif": SarifReporter,
     }